+++ lib/schleuder/plugin_runners/request_plugins_runner.rb
@@ -4,11 +4,19 @@ module Schleuder
       extend Base
 
       def self.notify_admins(keyword, arguments, response)
-        explanation = I18n.t('plugins.keyword_admin_notify_request',
+        if arguments.blank?
+          explanation = I18n.t('plugins.keyword_admin_notify_request_without_arguments',
+                                signer: @mail.signer,
+                                keyword: keyword
+                            )
+        else
+          explanation = I18n.t('plugins.keyword_admin_notify_request',
                                 signer: @mail.signer,
                                 keyword: keyword,
-                                arguments: arguments
+                                arguments: arguments.join(' ')
                             )
+        end
+        response = response.join("\n\n")
         msg = "#{explanation}\n\n#{response}"
         @list.logger.notify_admin(msg, nil, 'Notice')
       end