+++ spec/schleuder/integration/cli_spec.rb
@@ -80,11 +80,56 @@ describe 'cli' do
 
       output = run_cli("migrate #{v2list_path}")
       list = Schleuder::List.by_recipient('v2list@example.org')
+      admins_emails = list.admins.map(&:email)
       subscription_emails = list.subscriptions.map(&:email)
 
       expect(output).not_to match('Error:')
 
+      expect(admins_emails).to eql(["schleuder2@example.org"])
+
+      expect(subscription_emails).to eql(["anotherone@example.org", "anyone@example.org", "bla@foo", "old@example.org", "schleuder2@example.org", "someone@example.org"])
+      expect(list.subscriptions.where(email: "anotherone@example.org").first.fingerprint).to eql('')
+      expect(list.subscriptions.where(email: "anyone@example.org").first.fingerprint).to     eql("C4D60F8833789C7CAA44496FD3FFA6613AB10ECE")
+      expect(list.subscriptions.where(email: "bla@foo").first.fingerprint).to                eql("87E65ED2081AE3D16BE4F0A5EBDBE899251F2412")
+      expect(list.subscriptions.where(email: "old@example.org").first.fingerprint).to        eql("6EE51D78FD0B33DE65CCF69D2104E20E20889F66")
+      expect(list.subscriptions.where(email: "schleuder2@example.org").first.fingerprint).to eql("C4D60F8833789C7CAA44496FD3FFA6613AB10ECE")
+      expect(list.subscriptions.where(email: "someone@example.org").first.fingerprint).to    eql('')
+    end
+
+    it "does not fail on duplicated v2 subscriptions" do
+      v2list_path = 'spec/fixtures/v2list_duplicate_members'
+
+      output = run_cli("migrate #{v2list_path}")
+      expect(output).not_to match('Error:')
+      list = Schleuder::List.by_recipient('v2list@example.org')
+      subscription_emails = list.subscriptions.map(&:email)
+
       expect(subscription_emails).to eq ['schleuder2@example.org']
+    end
+    it "respects non delivery status of admins" do
+      v2list_path = 'spec/fixtures/v2list_admin_non_delivery'
+
+      output = run_cli("migrate #{v2list_path}")
+      expect(output).not_to match('Error:')
+      list = Schleuder::List.by_recipient('v2list@example.org')
+      subscriptions = list.subscriptions
+      expect(subscriptions.find{|s| s.email == 'schleuder2@example.org' }.delivery_enabled).to eq false
+      subscription_emails = subscriptions.map(&:email)
+      expect(subscription_emails.sort).to eq(['schleuder2@example.org',
+                                              'schleuder2-member@example.org'].sort)
+    end
+    it "does not fail on admin without key" do
+      v2list_path = 'spec/fixtures/v2list_admin_without_key'
+
+      output = run_cli("migrate #{v2list_path}")
+      expect(output).not_to match('Error:')
+
+      list = Schleuder::List.by_recipient('v2list@example.org')
+      admin_emails = list.admins.map(&:email)
+
+
+      expect(admin_emails.sort).to eq( ['schleuder2@example.org',
+                                        'schleuder2-nokey@example.org' ].sort)
     end
   end
 
@@ -101,9 +146,29 @@ describe 'cli' do
       mail = Mail::TestMailer.deliveries.first
 
       expect(Mail::TestMailer.deliveries.length).to eq 1
-      expect(mail.first_plaintext_part.body.to_s).to eql("Refreshing all keys from the keyring of list #{list.email} resulted in this:\n\nKey 59C71FB38AEE22E091C78259D06350440F759BD3 was updated (unchanged).\nKey 98769E8A1091F36BD88403ECF71A3F8412D83889 was updated (new signatures).\nKey 6EE51D78FD0B33DE65CCF69D2104E20E20889F66 was updated (new user-IDs, new signatures).")
+      expect(mail.first_plaintext_part.body.to_s).to eql("Refreshing all keys from the keyring of list #{list.email} resulted in this:\n\nKey 98769E8A1091F36BD88403ECF71A3F8412D83889 was updated (new signatures).\nKey 6EE51D78FD0B33DE65CCF69D2104E20E20889F66 was updated (new user-IDs, new signatures).")
 
       teardown_list_and_mailer(list)
+    end
+    it 'updates keys from the keyserver for only a specific list' do
+      list1 = create(:list)
+      list2 = create(:list)
+      [list1,list2].each do |list|
+        list.subscribe("admin@example.org", nil, true)
+        list.import_key(File.read("spec/fixtures/expired_key.txt"))
+        list.import_key(File.read("spec/fixtures/olduid_key.txt"))
+      end
+
+      with_sks_mock do
+        Cli.new.refresh_keys list1.email
+      end
+      mail = Mail::TestMailer.deliveries.first
+
+      expect(Mail::TestMailer.deliveries.length).to eq 1
+      expect(mail.first_plaintext_part.body.to_s).to eql("Refreshing all keys from the keyring of list #{list1.email} resulted in this:\n\nKey 98769E8A1091F36BD88403ECF71A3F8412D83889 was updated (new signatures).\nKey 6EE51D78FD0B33DE65CCF69D2104E20E20889F66 was updated (new user-IDs, new signatures).")
+
+      teardown_list_and_mailer(list1)
+      teardown_list_and_mailer(list2)
     end
 
     it 'reports errors from refreshing keys' do
@@ -122,6 +187,65 @@ describe 'cli' do
         # The wording differs slightly among versions.
         expect(mail.to_s).to match(/gpgkeys: .* error .* connect/)
       end
+
+      teardown_list_and_mailer(list)
+    end
+  end
+  context '#pin_keys' do
+    it 'pins fingerprints on not yet set keys' do
+      list = create(:list)
+      list.subscribe("admin@example.org", nil, true)
+      list.subscribe("schleuder2@example.org", nil, false)
+      list.import_key(File.read('spec/fixtures/example_key.txt'))
+      expect(list.subscriptions_without_fingerprint.size).to eq 2
+
+      Cli.new.pin_keys
+
+      expect(list.subscriptions_without_fingerprint.size).to eq 1
+      expect(list.subscriptions_without_fingerprint.collect(&:email)).to eq ['admin@example.org']
+
+      mail = Mail::TestMailer.deliveries.first
+
+      expect(Mail::TestMailer.deliveries.length).to eq 1
+      expect(mail.first_plaintext_part.body.to_s).to eql("While checking all subscriptions of list #{list.email} we were pinning a matching key for the following subscriptions:\n\nschleuder2@example.org: C4D60F8833789C7CAA44496FD3FFA6613AB10ECE")
+
+      teardown_list_and_mailer(list)
+    end
+    it 'only works on the specific list' do
+      list1 = create(:list)
+      list2 = create(:list)
+      [list1,list2].each do |list|
+        list.subscribe("admin@example.org", nil, true)
+        list.subscribe("schleuder2@example.org", nil, false)
+        list.import_key(File.read('spec/fixtures/example_key.txt'))
+        expect(list.subscriptions_without_fingerprint.size).to eq 2
+      end
+
+      Cli.new.pin_keys list1.email
+
+      expect(list1.subscriptions_without_fingerprint.size).to eq 1
+      expect(list1.subscriptions_without_fingerprint.collect(&:email)).to eq ['admin@example.org']
+      expect(list2.subscriptions_without_fingerprint.size).to eq 2
+
+      mail = Mail::TestMailer.deliveries.first
+
+      expect(Mail::TestMailer.deliveries.length).to eq 1
+      expect(mail.first_plaintext_part.body.to_s).to eql("While checking all subscriptions of list #{list1.email} we were pinning a matching key for the following subscriptions:\n\nschleuder2@example.org: C4D60F8833789C7CAA44496FD3FFA6613AB10ECE")
+
+      teardown_list_and_mailer(list1)
+      teardown_list_and_mailer(list2)
+    end
+
+    it 'does not report anything if nothing was done' do
+      list = create(:list)
+      list.subscribe("admin@example.org", nil, true)
+      list.subscribe("schleuder2@example.org", nil, false)
+      expect(list.subscriptions_without_fingerprint.size).to eq 2
+
+      Cli.new.pin_keys
+
+      expect(list.subscriptions_without_fingerprint.size).to eq 2
+      expect(Mail::TestMailer.deliveries.empty?).to eq true
 
       teardown_list_and_mailer(list)
     end